fix: audiopanorama panorama property requires f32 (gfloat), not f64
The GstAudioPanorama element's 'panorama' property is typed as gfloat.
Passing a Rust f64 (gdouble) caused a glib type validation panic when
any clip was loaded in the program player.

Cast clip.pan / pan argument to f32 before set_property call.

+++ src/media/program_player.rs
@@ -242,7 +242,7 @@ impl ProgramPlayer {
     pub fn update_current_audio(&mut self, volume: f64, pan: f64) {
         self.pipeline.set_property("volume", volume.clamp(0.0, 2.0));
         if let Some(ref ap) = self.audiopanorama {
-            ap.set_property("panorama", pan.clamp(-1.0, 1.0));
+            ap.set_property("panorama", (pan as f32).clamp(-1.0, 1.0));
         }
     }
 
@@ -272,7 +272,7 @@ impl ProgramPlayer {
         // Apply per-clip audio volume and pan
         self.pipeline.set_property("volume", clip.volume.clamp(0.0, 2.0));
         if let Some(ref ap) = self.audiopanorama {
-            ap.set_property("panorama", clip.pan.clamp(-1.0, 1.0));
+            ap.set_property("panorama", (clip.pan as f32).clamp(-1.0, 1.0));
         }
 
         if self.current_idx == Some(idx) {
